# src/production-service/openapi/production-service.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Doorstar Production Service API",
    "version": "0.1.0",
    "description": "Contract for the live Doorstar Üzemi Tábla REST service. The Planning C# service is a separate, future contract."
  },
  "servers": [{ "url": "/", "description": "Same-origin nginx deployment" }],
...
  "paths": {
...
    "/api/production/production-orders/{projectKey}/revisions/{revision}/readiness": {
      "get": { "tags": ["Readiness"], "operationId": "getOrderRevisionReadiness", "description": "Server-authoritative, read-only projection for one exact order revision, evaluated in one repeatable-read snapshot. A historical revision remains readable but can never be projected as current. Missing planning/work-package authority gives PRODUCTION_RELEASE the explicit NOT_AVAILABLE state.", "parameters": [{ "name": "projectKey", "in": "path", "required": true, "schema": { "type": "string", "minLength": 1 } }, { "name": "revision", "in": "path", "required": true, "schema": { "type": "integer", "minimum": 1 } }, { "$ref": "#/components/parameters/ReadModelRole" }], "responses": { "200": { "description": "Exact-revision readiness projection", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrderRevisionReadiness" } } } }, "400": { "$ref": "#/components/responses/ReadModelError" }, "403": { "$ref": "#/components/responses/ReadModelError" }, "404": { "$ref": "#/components/responses/ReadModelError" }, "409": { "$ref": "#/components/responses/ReadModelError" } } }
    },
    "/api/production/projects/{projectKey}/workflow": {
      "get": { "tags": ["Readiness"], "operationId": "getProjectWorkflow", "description": "Read-only latest-revision workflow projection evaluated in the same repeatable-read snapshot as its exact-revision readiness. Order, component and operation gates reuse exact-revision authority; planning, immutable work package, 6-stage runtime and handover stay CONTRACT_REQUIRED and never infer authority from legacy Tasks.", "parameters": [{ "name": "projectKey", "in": "path", "required": true, "schema": { "type": "string", "minLength": 1 } }, { "$ref": "#/components/parameters/ReadModelRole" }], "responses": { "200": { "description": "Authoritative and explicitly unavailable project workflow gates", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectWorkflow" } } } }, "400": { "$ref": "#/components/responses/ReadModelError" }, "403": { "$ref": "#/components/responses/ReadModelError" }, "404": { "$ref": "#/components/responses/ReadModelError" }, "409": { "$ref": "#/components/responses/ReadModelError" } } }
    },
...
    "/api/production/import-runs/{importRunId}/apply-manufactured-items": {
      "post": { "tags": ["Imports"], "operationId": "applyManufacturedItemCandidates", "parameters": [{ "$ref": "#/components/parameters/Role" }, { "name": "importRunId", "in": "path", "required": true, "schema": { "type": "string" } }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApplyManufacturedItemCandidates" } } } }, "responses": { "200": { "description": "Selected READY candidates applied idempotently to the imported test DRAFT" }, "400": { "$ref": "#/components/responses/InvalidRequest" }, "403": { "$ref": "#/components/responses/ManagerRequired" }, "404": { "description": "Import run not found" }, "409": { "description": "Test-schema, fingerprint, revision, candidate state or payload conflict" } } }
    },
...
  },
  "components": {
    "parameters": {
...
      "Role": { "name": "X-Role", "in": "header", "required": false, "description": "Temporary login-less role hint. `vezeto` is required by manager-only mutations; this is not authentication.", "schema": { "type": "string", "default": "vezeto" } },
      "ReadModelRole": { "name": "X-Role", "in": "header", "required": false, "description": "Temporary read hint only. Responses and authority owners always use canonical Doorstar roles; legacy `vezeto`/`allomas` are mapped to administrator/shop_floor. Omission is read-only reader.", "schema": { "type": "string", "enum": ["sales", "technical_preparation", "order_approver", "production_planner", "shop_floor", "installer", "warehouse_dispatch", "administrator", "reader", "vezeto", "allomas"], "default": "reader" } },
...
    },
    "responses": {
      "InvalidRequest": { "description": "Zod validation failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationError" } } } },
...
      "ManagerRequired": { "description": "Manager role required", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
      ,"ReadModelError": { "description": "Stable readiness/workflow read-model error envelope", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReadModelErrorEnvelope" } } } }
    },
    "schemas": {
...
      "CanonicalDoorstarRole": { "type": "string", "enum": ["sales", "technical_preparation", "order_approver", "production_planner", "shop_floor", "installer", "warehouse_dispatch", "administrator", "reader"] },
      "ReadModelErrorBody": { "type": "object", "required": ["code", "message", "details", "requestId"], "properties": { "code": { "type": "string", "enum": ["invalid_revision", "role_not_authorized", "project_not_found", "order_revision_not_found", "readiness_projection_conflict"] }, "message": { "type": "string" }, "details": { "type": "object", "additionalProperties": true }, "requestId": { "type": "string" } }, "additionalProperties": false },
      "ReadModelErrorEnvelope": { "type": "object", "required": ["error"], "properties": { "error": { "$ref": "#/components/schemas/ReadModelErrorBody" } }, "additionalProperties": false },
      "ReadinessEntity": { "type": "object", "required": ["kind", "id", "href"], "properties": { "kind": { "type": "string" }, "id": { "type": "string" }, "href": { "type": "string" } }, "additionalProperties": false },
      "ReadinessBlocker": { "type": "object", "required": ["code", "message", "ownerRole", "entity", "detail"], "properties": { "code": { "type": "string" }, "message": { "type": "string" }, "ownerRole": { "$ref": "#/components/schemas/CanonicalDoorstarRole" }, "entity": { "oneOf": [{ "$ref": "#/components/schemas/ReadinessEntity" }, { "type": "null" }] }, "detail": { "type": "object", "additionalProperties": true } }, "additionalProperties": false },
      "ReadinessAllowedActionCode": { "type": "string", "enum": ["REQUEST_ORDER_REVIEW", "APPROVE_ORDER_REVISION", "CREATE_COMPONENT_SNAPSHOT", "VERIFY_COMPONENT_SNAPSHOT", "REJECT_COMPONENT_SNAPSHOT", "CREATE_OPERATION_PLAN_SNAPSHOT", "VERIFY_OPERATION_PLAN", "REJECT_OPERATION_PLAN"] },
      "ReadinessAllowedAction": { "type": "object", "required": ["code", "method", "href", "ownerRoles", "targetEntityId"], "properties": { "code": { "$ref": "#/components/schemas/ReadinessAllowedActionCode" }, "method": { "type": "string", "enum": ["POST", "PATCH"] }, "href": { "type": "string" }, "ownerRoles": { "type": "array", "minItems": 1, "uniqueItems": true, "items": { "$ref": "#/components/schemas/CanonicalDoorstarRole" } }, "targetEntityId": { "type": "string" } }, "additionalProperties": false },
      "ReadinessNextAction": { "oneOf": [{ "type": "object", "required": ["kind", "action"], "properties": { "kind": { "const": "ACTION" }, "action": { "$ref": "#/components/schemas/ReadinessAllowedAction" } }, "additionalProperties": false }, { "type": "object", "required": ["kind", "blockerCode", "ownerRole", "href"], "properties": { "kind": { "const": "BLOCKED" }, "blockerCode": { "type": "string" }, "ownerRole": { "$ref": "#/components/schemas/CanonicalDoorstarRole" }, "href": { "type": ["string", "null"] } }, "additionalProperties": false }, { "type": "object", "required": ["kind", "message"], "properties": { "kind": { "const": "COMPLETE" }, "message": { "type": "string" } }, "additionalProperties": false }] },
      "RevisionReadinessContentHash": { "type": "object", "required": ["value", "schemaVersion", "verification", "auditId"], "properties": { "value": { "type": "string", "pattern": "^[a-f0-9]{64}$" }, "schemaVersion": { "type": "integer", "enum": [1, 2, 3] }, "verification": { "type": "string", "enum": ["UNAPPROVED_CURRENT", "VERIFIED", "MISMATCH", "AUDIT_MISSING"] }, "auditId": { "type": ["string", "null"] } }, "additionalProperties": false },
      "ExactRevisionReadinessIdentity": { "type": "object", "required": ["id", "number", "isLatest", "latestRevisionNumber", "status", "intakeStage", "updatedAt", "contentHash"], "properties": { "id": { "type": "string" }, "number": { "type": "integer", "minimum": 1 }, "isLatest": { "type": "boolean" }, "latestRevisionNumber": { "type": "integer", "minimum": 1 }, "status": { "type": "string", "enum": ["DRAFT", "REVIEW", "APPROVED", "SUPERSEDED"] }, "intakeStage": { "type": "string", "enum": ["SALES_DRAFT", "SALES_DOCUMENTS_RECEIVED", "SURVEY_PENDING", "SURVEY_COMPLETED", "SURVEY_EXCEPTION_REVIEW", "TECHNICAL_PREPARATION"] }, "updatedAt": { "type": "string", "format": "date-time" }, "contentHash": { "$ref": "#/components/schemas/RevisionReadinessContentHash" } }, "additionalProperties": false },
      "SurveyReadinessDetails": { "type": "object", "required": ["kind", "positionCount", "surveyDocumentRequired", "positionsMissingFields", "positionIdsMissingSurveyDocumentLink", "positionEvidence"], "properties": { "kind": { "const": "SURVEY" }, "positionCount": { "type": "integer", "minimum": 0 }, "surveyDocumentRequired": { "type": "boolean" }, "positionsMissingFields": { "type": "array", "items": { "$ref": "#/components/schemas/SurveyPositionMissingFields" } }, "positionIdsMissingSurveyDocumentLink": { "type": "array", "items": { "type": "string" } }, "positionEvidence": { "$ref": "#/components/schemas/PositionEvidenceGateDetails" } }, "additionalProperties": false },
      "PositionEvidenceReadinessDetails": { "type": "object", "required": ["kind", "totalEvidence", "resolvedEvidence", "unresolvedEvidence", "rejectedEvidence", "blockerEvidenceIds"], "properties": { "kind": { "const": "POSITION_EVIDENCE" }, "totalEvidence": { "type": "integer", "minimum": 0 }, "resolvedEvidence": { "type": "integer", "minimum": 0 }, "unresolvedEvidence": { "type": "integer", "minimum": 0 }, "rejectedEvidence": { "type": "integer", "minimum": 0 }, "blockerEvidenceIds": { "type": "array", "items": { "type": "string" } } }, "additionalProperties": false },
      "DocumentReadinessDetails": { "type": "object", "required": ["kind", "total", "currentVersionIds", "supersededVersionIds", "staleLinkedDocumentVersionIds"], "properties": { "kind": { "const": "DOCUMENTS" }, "total": { "type": "integer", "minimum": 0 }, "currentVersionIds": { "type": "array", "items": { "type": "string" } }, "supersededVersionIds": { "type": "array", "items": { "type": "string" } }, "staleLinkedDocumentVersionIds": { "type": "array", "items": { "type": "string" } } }, "additionalProperties": false },
      "SourceItemReadinessDetails": { "type": "object", "required": ["kind", "total", "ready", "unresolved", "blockerItemIds"], "properties": { "kind": { "type": "string", "enum": ["MANUFACTURED_ITEMS", "SUPPLEMENTARY_ITEMS"] }, "total": { "type": "integer", "minimum": 0 }, "ready": { "type": "integer", "minimum": 0 }, "unresolved": { "type": "integer", "minimum": 0 }, "blockerItemIds": { "type": "array", "items": { "type": "string" } } }, "additionalProperties": false },
      "OrderReviewReadinessDetails": { "type": "object", "required": ["kind", "status", "intakeStage", "sharedPredicateReady", "contentHash", "openFeedbackCount", "feedbackIsApprovalBlocker"], "properties": { "kind": { "const": "ORDER_REVIEW" }, "status": { "type": "string", "enum": ["DRAFT", "REVIEW", "APPROVED", "SUPERSEDED"] }, "intakeStage": { "type": "string" }, "sharedPredicateReady": { "type": "boolean" }, "contentHash": { "$ref": "#/components/schemas/RevisionReadinessContentHash" }, "openFeedbackCount": { "type": "integer", "minimum": 0 }, "feedbackIsApprovalBlocker": { "const": false } }, "additionalProperties": false },
      "SnapshotReadinessDetails": { "type": "object", "required": ["kind", "snapshotCount", "selectedSnapshotId", "selectedState", "currentVerifiedSnapshotId", "current", "verified"], "properties": { "kind": { "type": "string", "enum": ["COMPONENT_SNAPSHOT", "OPERATION_PLAN"] }, "snapshotCount": { "type": "integer", "minimum": 0 }, "selectedSnapshotId": { "type": ["string", "null"] }, "selectedState": { "type": ["string", "null"], "enum": ["REVIEW", "VERIFIED", "REJECTED", null] }, "currentVerifiedSnapshotId": { "type": ["string", "null"] }, "current": { "type": "boolean" }, "verified": { "type": "boolean" }, "createReadiness": { "$ref": "#/components/schemas/OperationPlanReadiness" } }, "additionalProperties": false },
      "ProductionReleaseReadinessDetails": { "type": "object", "required": ["kind", "authority", "orderApproved", "componentVerified", "operationVerified", "planningProposalAvailable", "issuedWorkPackageAvailable"], "properties": { "kind": { "const": "PRODUCTION_RELEASE" }, "authority": { "const": "NOT_AVAILABLE" }, "orderApproved": { "type": "boolean" }, "componentVerified": { "type": "boolean" }, "operationVerified": { "type": "boolean" }, "planningProposalAvailable": { "const": false }, "issuedWorkPackageAvailable": { "const": false } }, "additionalProperties": false },
      "ReadinessGateDetails": { "oneOf": [{ "$ref": "#/components/schemas/SurveyReadinessDetails" }, { "$ref": "#/components/schemas/PositionEvidenceReadinessDetails" }, { "$ref": "#/components/schemas/DocumentReadinessDetails" }, { "$ref": "#/components/schemas/SourceItemReadinessDetails" }, { "$ref": "#/components/schemas/OrderReviewReadinessDetails" }, { "$ref": "#/components/schemas/SnapshotReadinessDetails" }, { "$ref": "#/components/schemas/ProductionReleaseReadinessDetails" }] },
      "ReadinessGateKey": { "type": "string", "description": "Response order is SURVEY, POSITION_EVIDENCE, DOCUMENTS, MANUFACTURED_ITEMS, SUPPLEMENTARY_ITEMS, ORDER_REVIEW, COMPONENT_SNAPSHOT, OPERATION_PLAN, PRODUCTION_RELEASE.", "enum": ["SURVEY", "POSITION_EVIDENCE", "DOCUMENTS", "MANUFACTURED_ITEMS", "SUPPLEMENTARY_ITEMS", "ORDER_REVIEW", "COMPONENT_SNAPSHOT", "OPERATION_PLAN", "PRODUCTION_RELEASE"] },
      "ReadinessGate": { "type": "object", "required": ["key", "state", "ready", "ownerRole", "detailsHref", "blockers", "allowedActions", "details"], "properties": { "key": { "$ref": "#/components/schemas/ReadinessGateKey" }, "state": { "type": "string", "enum": ["READY", "BLOCKED", "NOT_AVAILABLE"] }, "ready": { "type": "boolean" }, "ownerRole": { "$ref": "#/components/schemas/CanonicalDoorstarRole" }, "detailsHref": { "type": ["string", "null"] }, "blockers": { "type": "array", "items": { "$ref": "#/components/schemas/ReadinessBlocker" } }, "allowedActions": { "type": "array", "items": { "$ref": "#/components/schemas/ReadinessAllowedAction" } }, "details": { "$ref": "#/components/schemas/ReadinessGateDetails" } }, "additionalProperties": false },
      "OrderRevisionReadiness": { "type": "object", "required": ["schemaVersion", "projectKey", "revision", "gates", "blockers", "allowedActions", "nextAction"], "properties": { "schemaVersion": { "const": "doorstar.order-revision-readiness/v1" }, "projectKey": { "type": "string" }, "revision": { "$ref": "#/components/schemas/ExactRevisionReadinessIdentity" }, "gates": { "type": "array", "minItems": 9, "maxItems": 9, "items": { "$ref": "#/components/schemas/ReadinessGate" } }, "blockers": { "type": "array", "items": { "$ref": "#/components/schemas/ReadinessBlocker" } }, "allowedActions": { "type": "array", "items": { "$ref": "#/components/schemas/ReadinessAllowedAction" } }, "nextAction": { "$ref": "#/components/schemas/ReadinessNextAction" } }, "additionalProperties": false },
      "ProjectWorkflowSource": { "type": "object", "required": ["kind", "id", "revision", "contentHash", "href"], "properties": { "kind": { "const": "ORDER_REVISION" }, "id": { "type": "string" }, "revision": { "type": "integer", "minimum": 1 }, "contentHash": { "type": "string", "pattern": "^[a-f0-9]{64}$" }, "href": { "type": "string" } }, "additionalProperties": false },
      "ProjectWorkflowGate": { "type": "object", "required": ["key", "state", "ownerRole", "source", "blockers", "allowedActions", "detailsHref"], "properties": { "key": { "type": "string", "enum": ["ORDER", "COMPONENTS", "OPERATIONS", "PLANNING", "WORK_PACKAGE", "PRODUCTION_6_STAGE", "HANDOVER"] }, "state": { "type": "string", "enum": ["READY", "BLOCKED", "NOT_AVAILABLE", "CONTRACT_REQUIRED"] }, "ownerRole": { "$ref": "#/components/schemas/CanonicalDoorstarRole" }, "source": { "$ref": "#/components/schemas/ProjectWorkflowSource" }, "blockers": { "type": "array", "items": { "$ref": "#/components/schemas/ReadinessBlocker" } }, "allowedActions": { "type": "array", "items": { "$ref": "#/components/schemas/ReadinessAllowedAction" } }, "detailsHref": { "type": ["string", "null"] } }, "additionalProperties": false },
      "ProjectWorkflow": { "type": "object", "required": ["schemaVersion", "projectKey", "revision", "currentGate", "gates", "blockers", "allowedActions", "nextAction"], "properties": { "schemaVersion": { "const": "doorstar.project-workflow/v1" }, "projectKey": { "type": "string" }, "revision": { "type": "object", "required": ["id", "number", "isLatest", "href"], "properties": { "id": { "type": "string" }, "number": { "type": "integer", "minimum": 1 }, "isLatest": { "type": "boolean" }, "href": { "type": "string" } }, "additionalProperties": false }, "currentGate": { "type": ["string", "null"], "enum": ["ORDER", "COMPONENTS", "OPERATIONS", "PLANNING", "WORK_PACKAGE", "PRODUCTION_6_STAGE", "HANDOVER", null] }, "gates": { "type": "array", "minItems": 7, "maxItems": 7, "items": { "$ref": "#/components/schemas/ProjectWorkflowGate" } }, "blockers": { "type": "array", "items": { "$ref": "#/components/schemas/ReadinessBlocker" } }, "allowedActions": { "type": "array", "items": { "$ref": "#/components/schemas/ReadinessAllowedAction" } }, "nextAction": { "$ref": "#/components/schemas/ReadinessNextAction" } }, "additionalProperties": false },
      "Error": { "type": "object", "required": ["error"], "properties": { "error": { "type": "string" } }, "additionalProperties": true },
      "ValidationError": { "allOf": [{ "$ref": "#/components/schemas/Error" }, { "type": "object", "required": ["details"], "properties": { "error": { "const": "invalid_request" }, "details": { "type": "object" } } }] },
...
      "ApplyManufacturedItemCandidates": { "type": "object", "required": ["orderRevisionId", "sourceFingerprint", "candidateIds", "confirmation"], "properties": { "orderRevisionId": { "type": "string", "minLength": 1 }, "sourceFingerprint": { "type": "string", "pattern": "^[a-fA-F0-9]{64}$" }, "candidateIds": { "type": "array", "minItems": 1, "maxItems": 200, "uniqueItems": true, "items": { "type": "string", "minLength": 1 } }, "confirmation": { "const": "APPLY_READY_MANUFACTURED_ITEMS" } }, "additionalProperties": false },
...
    }
  }
}